Move old button to deprecated
