+++ src/__tests__/AnchoredOverlay.test.tsx
@@ -4,7 +4,8 @@ import {behavesAsComponent, checkExports} from '../utils/testing'
 import {render as HTMLRender, cleanup, fireEvent} from '@testing-library/react'
 import {axe, toHaveNoViolations} from 'jest-axe'
 import 'babel-polyfill'
-import {Button, SSRProvider} from '../index'
+import {SSRProvider} from '../index'
+import {Button} from '../deprecated'
 import theme from '../theme'
 import BaseStyles from '../BaseStyles'
 import {ThemeProvider} from '../ThemeProvider'
