+++ src/__tests__/ConfirmationDialog.test.tsx
@@ -6,7 +6,7 @@ import React, {useCallback, useRef, useState} from 'react'
 import {ActionMenu} from '../deprecated/ActionMenu'
 import BaseStyles from '../BaseStyles'
 import Box from '../Box'
-import Button from '../Button/Button'
+import Button from '../deprecated/Button/Button'
 import {ConfirmationDialog, useConfirm} from '../Dialog/ConfirmationDialog'
 import theme from '../theme'
 import {ThemeProvider} from '../ThemeProvider'
