+++ src/__tests__/Details.test.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
-import {Details, useDetails, Button, ButtonPrimary, Box} from '..'
-import {ButtonProps} from '../Button/Button'
+import {Details, useDetails, Box} from '..'
+import {Button, ButtonPrimary} from '../deprecated'
+import {ButtonProps} from '../deprecated/Button/Button'
 import {mount, behavesAsComponent, checkExports} from '../utils/testing'
 import {render as HTMLRender, cleanup} from '@testing-library/react'
 import {axe, toHaveNoViolations} from 'jest-axe'
