+++ src/deprecated/Button/Button.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components'
-import {get} from '../constants'
-import sx, {SxProp} from '../sx'
-import {ComponentProps} from '../utils/types'
+import {get} from '../../constants'
+import sx, {SxProp} from '../../sx'
+import {ComponentProps} from '../../utils/types'
 import ButtonBase, {ButtonBaseProps} from './ButtonBase'
 
 const Button = styled(ButtonBase)<ButtonBaseProps & SxProp>`
