+++ src/deprecated/Button/ButtonClose.tsx
@@ -1,9 +1,9 @@
 import {XIcon} from '@primer/octicons-react'
 import React, {forwardRef} from 'react'
 import styled from 'styled-components'
-import {get} from '../constants'
-import sx, {SxProp} from '../sx'
-import {ComponentProps} from '../utils/types'
+import {get} from '../../constants'
+import sx, {SxProp} from '../../sx'
+import {ComponentProps} from '../../utils/types'
 
 const StyledButton = styled.button<SxProp>`
   border: none;
