+++ src/deprecated/Button/index.ts
@@ -2,8 +2,6 @@ export {default} from './Button'
 export type {ButtonProps} from './Button'
 export {default as ButtonDanger} from './ButtonDanger'
 export type {ButtonDangerProps} from './ButtonDanger'
-export {default as ButtonGroup} from './ButtonGroup'
-export type {ButtonGroupProps} from './ButtonGroup'
 export {default as ButtonOutline} from './ButtonOutline'
 export type {ButtonOutlineProps} from './ButtonOutline'
 export {default as ButtonPrimary} from './ButtonPrimary'
