+++ src/index.ts
@@ -52,26 +52,8 @@ export {default as BranchName} from './BranchName'
 export type {BranchNameProps} from './BranchName'
 export {default as Breadcrumbs, Breadcrumb} from './Breadcrumbs'
 export type {BreadcrumbsProps, BreadcrumbsItemProps, BreadcrumbProps, BreadcrumbItemProps} from './Breadcrumbs'
-export {
-  default as Button,
-  ButtonDanger,
-  ButtonOutline,
-  ButtonPrimary,
-  ButtonInvisible,
-  ButtonTableList,
-  ButtonClose,
-  ButtonGroup
-} from './Button'
-export type {
-  ButtonProps,
-  ButtonDangerProps,
-  ButtonOutlineProps,
-  ButtonPrimaryProps,
-  ButtonInvisibleProps,
-  ButtonTableListProps,
-  ButtonCloseProps,
-  ButtonGroupProps
-} from './Button'
+export {default as ButtonGroup} from './ButtonGroup'
+export type {ButtonGroupProps} from './ButtonGroup'
 export {default as Caret} from './Caret'
 export type {CaretProps} from './Caret'
 export {default as ChoiceInputField} from './ChoiceInputField'
