+++ src/stories/Button.stories.tsx
@@ -2,19 +2,17 @@ import React from 'react'
 import {Meta} from '@storybook/react'
 
 import {
-  BaseStyles,
   Button,
   ButtonClose,
   ButtonDanger,
-  ButtonGroup,
   ButtonInvisible,
   ButtonOutline,
   ButtonPrimary,
-  ButtonTableList,
-  ThemeProvider
-} from '..'
+  ButtonTableList
+} from '../deprecated'
+import {BaseStyles, ButtonGroup, ThemeProvider} from '..'
 import {ButtonStyleProps} from 'styled-system'
-import {ButtonBaseProps} from '../Button/ButtonBase'
+import {ButtonBaseProps} from '../deprecated/Button/ButtonBase'
 type StrictButtonStyleProps = ButtonStyleProps & {variant: ButtonBaseProps['variant']}
 
 export default {
