+++ src/stories/ConfirmationDialog.stories.tsx
@@ -1,7 +1,7 @@
 import React, {useState, useRef, useCallback} from 'react'
 import {Meta} from '@storybook/react'
-
-import {BaseStyles, Button, Box, ThemeProvider, useTheme} from '..'
+import {BaseStyles, Box, ThemeProvider, useTheme} from '..'
+import {Button} from '../deprecated'
 import {ConfirmationDialog, useConfirm} from '../Dialog/ConfirmationDialog'
 import {ActionMenu} from '../deprecated/ActionMenu'
