+++ src/stories/deprecated/ActionMenu.stories.tsx
@@ -15,8 +15,7 @@ import React, {useCallback, useState, useRef} from 'react'
 import styled from 'styled-components'
 import {ThemeProvider} from '../..'
 import Link, {LinkProps} from '../../Link'
-import Button from '../../Button'
-import {ActionMenu, ActionMenuProps, ActionList, DropdownButton} from '../../deprecated'
+import {ActionMenu, ActionMenuProps, ActionList, DropdownButton, Button} from '../../deprecated'
 import {ItemProps} from '../../deprecated/ActionList'
 import BaseStyles from '../../BaseStyles'
