+++ src/stories/useAnchoredPosition.stories.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import {Meta} from '@storybook/react'
-import {BaseStyles, Box, ButtonPrimary, ThemeProvider} from '..'
+import {BaseStyles, Box, ThemeProvider} from '..'
+import {ButtonPrimary} from '../deprecated'
 import {useAnchoredPosition} from '../hooks'
 import styled from 'styled-components'
 import {get} from '../constants'
